requirements.txt
pyTelegramBotAPI==4.14.0
gspread==5.12.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-cloud-texttospeech==2.16.3
requests==2.31.0
cairosvg==2.7.0
protobuf==4.25.8